refactor : index, appinit, app
