+++ public/index.css
@@ -1,9 +1,3 @@
 #root:not(:empty) ~ #loading {
   display: none;
 }
-
-#loading {
-  background-color: #f9f9f9;
-  width: 100vw;
-  height: 100vh;
-}
