+++ public/index.html
@@ -6,7 +6,7 @@
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="Web site created using create-react-app"
+      content="유튜브 영상의 원하는 부분만을 정해서 플레이리스트를 만들 수 있습니다."
     />
 
 
@@ -77,7 +77,7 @@
       css 파일을 연동하는 시간도 아끼기 위해 inline style도 사용함
       번들 다운로드 후, 내부 api 작업을 위한 로딩은 #root 안에서 해결
     -->
-    <div id="loading">
+    <div id="loading" style="background-color:#f9f9f9; width:100vw; height:100vh;">
         <img id="loadingElement" src="%PUBLIC_URL%/loading.png" style="position:absolute; width:100px; top:50%; left:50%;"/>
     </div>
   </body>
