+++ src/lib/hooks/useScreenSize.ts
@@ -1,3 +1,12 @@
+/**
+ * window 사이즈 너비의 변화를 트래킹하고,
+ * 어느 범위에 속한 screenSize인지 확인 후 사이즈를 업데이트함.
+ *
+ * 주의점 :
+ *  resize 이벤트를 트래킹하는 hook이기에, 앱의 최상단인 appInit에서 딱 한번 사용하고,
+ *  screenSize 정보는 리덕스에 저장하는 식으로하여, 이벤트를 여러개 등록하지 않도록 한다.
+ */
+
 import { useEffect, useState } from 'react'
 import * as Constants from '../constants'
